move roomt creation to separate component
