+++ package.json
@@ -1,6 +1,6 @@
 {
     "name": "Meetings",
-    "version": "2.23.0",
+    "version": "2.24.0",
     "description": "",
     "main": "index.js",
     "scripts": {
