+++ plugin.manifest
@@ -1,7 +1,7 @@
 pluginname=Meetings
 pluginclassname=MeetingPlugin
 origin=elan-ev
-version=2.23
+version=2.24
 studipMinVersion=4.0
 studipMaxVersion=4.6.99
 description=Virtueller Raum, mit dem Live-Online-Treffen, Veranstaltungen und Videokonferenzen durchgeführt werden können.
